feat(policy-api):[#639] Extract pagination logic to own service in order to avoid PolicyStoreService becoming a god-class
